Move ST_PlanarGraph test to its own class.
