+++ src/test/java/org/gdms/gdmstopology/function/GraphBuilderTest.java
@@ -51,101 +51,6 @@ import org.orbisgis.progress.NullProgressMonitor;
 public class GraphBuilderTest extends TopologySetupTest {
 
     /**
-     * A test to validate the planar graph method with two polygons.
-     *
-     * @throws Exception
-     */
-    @Test
-    public void testST_PlanarGraph() throws Exception {
-        //Input datasource
-        final MemoryDataSetDriver driver_src = new MemoryDataSetDriver(
-                new String[]{
-                    "the_geom",
-                    "id"
-                },
-                new Type[]{
-                    TypeFactory.createType(Type.GEOMETRY),
-                    TypeFactory.createType(Type.INT)
-                });
-
-        driver_src.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("POLYGON( ( 69 152, 69 293, 221 293, 221 152, 69 152 ))")),
-                    ValueFactory.createValue(1)
-                });
-        driver_src.addValues(
-                new Value[]{
-                    ValueFactory.createValue(wktReader.read("POLYGON( ( 221 152, 221 293, 390 293, 390 152, 221 152 ))")),
-                    ValueFactory.createValue(2)
-                });
-
-
-        ST_PlanarGraph st_PlanarGraph = new ST_PlanarGraph();
-        DataSet[] tables = new DataSet[]{driver_src};
-        st_PlanarGraph.evaluate(
-                dsf,
-                tables,
-                new Value[]{
-                    ValueFactory.createValue("output")
-                },
-                new NullProgressMonitor());
-
-
-        DataSource dsResult_polygons = dsf.getDataSource("output.polygons");
-
-
-        dsResult_polygons.open();
-        //The planar graph returns 2 polygons
-        assertTrue(dsResult_polygons.getRowCount() == 2);
-
-        dsResult_polygons.close();
-        DataSource dsResult_edges = dsf.getDataSource("output.edges");
-
-        dsResult_edges.open();
-        //The planar graph returns 3 lines
-        assertTrue(dsResult_edges.getRowCount() == 3);
-
-        //Test the gid values for the 3 lines                
-        Value[] row0 = dsResult_edges.getRow(0);
-        //This row corresponds to the shared segment between geometry from row0 and row1
-        Value[] row1 = dsResult_edges.getRow(1);
-        Value[] row2 = dsResult_edges.getRow(2);
-
-        int gidRight = row1[4].getAsInt();
-        int gidLeft = row1[5].getAsInt();
-
-        //Check the left and right gids
-        //if the gids are not equal do a flip test, otherwise the topology is not valid.
-        if (gidLeft == 1 && gidRight == 2) {
-            //Left gid
-            assertTrue(row0[5].getAsInt() == -1);
-            assertTrue(row2[5].getAsInt() == -1);
-            //Right gid
-            assertTrue(row0[4].getAsInt() == 1);
-            assertTrue(row2[4].getAsInt() == 2);
-        } else if (gidLeft == 2 && gidRight == 1) {
-            //Left gid
-            assertTrue(row0[5].getAsInt() == -1);
-            assertTrue(row2[5].getAsInt() == -1);
-            //Right gid
-            assertTrue(row0[4].getAsInt() == 2);
-            assertTrue(row2[4].getAsInt() == 1);
-        } else {
-            assertTrue("The topology is not valid", false);
-        }
-
-        //Check the start and end gids
-        assertTrue(row0[2].getAsInt() == 1);
-        assertTrue(row0[3].getAsInt() == 2);
-        assertTrue(row1[2].getAsInt() == 1);
-        assertTrue(row1[3].getAsInt() == 2);
-        assertTrue(row2[2].getAsInt() == 2);
-        assertTrue(row2[3].getAsInt() == 1);
-
-        dsResult_edges.close();
-    }
-
-    /**
      * A test to validate the network graph method
      *
      * @throws Exception
